rename src to module name
This keeps poetry happy
